When a component's "main" is an array, use the first item in the array
`bower.json` defines the `main` option as (validly) being either a string or an array. debowerify previously assumed that `main` was always a string, so valid Bower components that chose to use an array instead would produce a `TypeError`. This patch has debowerify test the content of `main`, and if it's an array it will instead use the first element of the array.