Improves error message emitted when there's a problem in debowerify
For some reason, the 'self' reference was previously often undefined in case of error, which meant the error shown to the user was along the lines of "undefined has no method 'self'" instead of the actual cause of problems. Simply referencing 'tr' directly (which is what every other reference to the stream in the file does anyway) seems to correct this.