Fixes offline bug
Best I could do with research brought me to conclude that the bower list
command never supported the `map` option. The side effect was that the
bower code was attempting to check for new versions of the packages the
first time debowerify tried to scan for dependencies. This caused the
promise to be rejected when your on a plane and have no internet access,
causing the `error` event to trigger, in turn causeing debowerify to
bail out and end the build. This means you could not build your apps
while on a plane or where you don't have online access.

Bower allows an offline option to prevent this. I changed the unused
`map` option to the now used `offline` option and the return value from
bower is the same.