Returning an error message if bower reports a dependency as missing
When debowerify is ran before a bower install, it returns an obscure error related to path.join (because bower has found the project in the json file but it was not installed). In this case bower returns module.missing as true, so I have added an exception to help the user into understanding why the process fail.