Introduce overrides for `main`
Debowerify doesn't really help if the bower package doesn't declare the
right files in the `main` section of their bower.json file. This patch
introduces the `overrides` section that allows the user declaring
dependencies to override the main section of the bower.json file of the
package they depend on.

This strategy was copied from `main-bower-files`:

  https://github.com/ck86/main-bower-files#options